+++ packages/sdk/src/Configuration/Fee.ts
@@ -1,7 +1,6 @@
 import * as Abis from "@enzymefinance/abis";
-import { Address, PublicClient, isAddressEqual } from "viem";
+import { Address, PublicClient } from "viem";
 import { Viem } from "../Utils.js";
-import { getInfo } from "./Fees/Performance.js";
 
 export {
   Action,
@@ -14,18 +13,18 @@ export {
   type SettleContinuousFeesParams,
 } from "../_internal/FeeManager.js";
 
-export async function getEnabledFees(
+export function getRecipient(
   client: PublicClient,
   args: Viem.ContractCallParameters<{
     comptrollerProxy: Address;
-    feeManager: Address;
+    fee: Address;
   }>,
 ) {
   return Viem.readContract(client, args, {
-    abi: Abis.IFeeManager,
-    functionName: "getEnabledFeesForFund",
+    abi: Abis.IFee,
+    functionName: "getRecipientForFund",
     args: [args.comptrollerProxy],
-    address: args.feeManager,
+    address: args.fee,
   });
 }
 
@@ -111,79 +110,4 @@ export async function getMlnValueAndBurnAmountForSharesBuyback(
   const mlnAmountToBurn = mlnValueOfBuyback / 2n;
 
   return { mlnAmountToBurn, mlnValue: mlnValueOfBuyback };
-}
-
-export async function getAccruedContinuousFees(
-  client: PublicClient,
-  args: Viem.ContractCallParameters<{
-    feeManager: Address;
-    unpermissionedActionsWrapper: Address;
-    managementFee: Address;
-    performanceFee: Address;
-    comptrollerProxy: Address;
-    vaultProxy: Address;
-  }>,
-) {
-  const continuousFees = await Viem.readContract(client, args, {
-    abi: Abis.IUnpermissionedActionsWrapper,
-    functionName: "getContinuousFeesForFund",
-    address: args.unpermissionedActionsWrapper,
-    args: [args.comptrollerProxy],
-  });
-
-  const hasManagementFee = continuousFees.some((fee) => isAddressEqual(fee, args.managementFee));
-  const hasPerformanceFee = continuousFees.some((fee) => isAddressEqual(fee, args.performanceFee));
-
-  let managementFeeSharesDue: bigint | undefined;
-
-  if (hasManagementFee) {
-    const {
-      result: [_, __, sharesDue],
-    } = await Viem.simulateContract(client, args, {
-      abi: Abis.IManagementFee,
-      functionName: "settle",
-      address: args.managementFee,
-      args: [args.comptrollerProxy, args.vaultProxy, 0, "0x", 0n],
-      account: args.feeManager,
-    });
-
-    managementFeeSharesDue = sharesDue;
-  }
-
-  let performanceFeeSharesDue: bigint | undefined;
-  let highWaterMark: bigint | undefined;
-
-  if (hasPerformanceFee) {
-    const { result: gav } = await Viem.simulateContract(client, args, {
-      abi: Abis.IComptrollerLib,
-      functionName: "calcGav",
-      address: args.comptrollerProxy,
-    });
-
-    const {
-      result: [_, __, sharesDue],
-    } = await Viem.simulateContract(client, args, {
-      abi: Abis.IPerformanceFee,
-      functionName: "settle",
-      address: args.managementFee,
-      args: [args.comptrollerProxy, args.vaultProxy, 0, "0x", gav],
-      account: args.feeManager,
-    });
-
-    performanceFeeSharesDue = sharesDue;
-
-    const performanceFeeInfo = await getInfo(client, {
-      performanceFee: args.performanceFee,
-      comptrollerProxy: args.comptrollerProxy,
-    });
-
-    highWaterMark = performanceFeeInfo.highWaterMark;
-  }
-
-  return {
-    continuousFees,
-    highWaterMark,
-    managementFeeSharesDue,
-    performanceFeeSharesDue,
-  };
 }
