+++ packages/sdk/src/Configuration/Policies.ts
@@ -1,7 +1,3 @@
-import * as Abis from "@enzymefinance/abis";
-import type { Address, PublicClient } from "viem";
-import { Viem } from "../Utils.js";
-
 export * as AllowedAdapterIncomingAssets from "./Policies/AllowedAdapterIncomingAssets.js";
 export * as AllowedAdapters from "./Policies/AllowedAdapters.js";
 export * as AllowedAdaptersPerManager from "./Policies/AllowedAdaptersPerManager.js";
@@ -13,51 +9,3 @@ export * as AllowedSharesTransferRecipients from "./Policies/AllowedSharesTransf
 export * as CumulativeSlippageTolerance from "./Policies/CumulativeSlippageTolerance.js";
 export * as MinAssetBalancesPostRedemption from "./Policies/MinAssetBalancesPostRedemption.js";
 export * as MinMaxInvestment from "./Policies/MinMaxInvestment.js";
-
-const getListIdsForFundAbi = {
-  inputs: [
-    {
-      internalType: "address",
-      name: "_comptrollerProxy",
-      type: "address",
-    },
-  ],
-  name: "getListIdsForFund",
-  outputs: [
-    {
-      internalType: "uint256[]",
-      name: "listIds_",
-      type: "uint256[]",
-    },
-  ],
-  stateMutability: "view",
-  type: "function",
-} as const;
-
-export function getListIds(
-  client: PublicClient,
-  args: Viem.ContractCallParameters<{
-    policyContract: Address;
-    comptrollerProxy: Address;
-  }>,
-) {
-  return Viem.readContract(client, args, {
-    abi: [getListIdsForFundAbi],
-    functionName: "getListIdsForFund",
-    args: [args.comptrollerProxy],
-    address: args.policyContract,
-  });
-}
-
-export function getIdentifier(
-  client: PublicClient,
-  args: Viem.ContractCallParameters<{
-    policy: Address;
-  }>,
-) {
-  return Viem.readContract(client, args, {
-    abi: Abis.IPolicy,
-    functionName: "identifier",
-    address: args.policy,
-  });
-}
